Crate resourceviews1_beta2

This documentation was generated from resourceviews crate version 0.0.1. The original source code can be found on github.

Features

Handle the following Resources with ease from the central hub ...

zone operations (get and list)

zone views (add resources, delete, get, get service, insert, list, list resources, remove resources and set service)

Everything else about the resourceviews v1_beta2 API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.zone_views().remove_resources(...).doit()
let r = hub.zone_views().add_resources(...).doit()
let r = hub.zone_views().delete(...).doit()
let r = hub.zone_operations().get(...).doit()
let r = hub.zone_views().insert(...).doit()
let r = hub.zone_views().set_service(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
resourceviews1_beta2 = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate resourceviews1_beta2;
use resourceviews1_beta2::ZoneViewsRemoveResourcesRequest;
use resourceviews1_beta2::cmn::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use resourceviews1_beta2::Resourceviews;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Resourceviews::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req: ZoneViewsRemoveResourcesRequest = Default::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.zone_views().remove_resources(&req, "project", "zone", "resourceView")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Macros

map!
This macro is advertised in the documentation, which is why we deliver it as well

Structs

Label
The Label to be applied to the resource views.

This type is not used in any activity, and only used as part of another schema.

ListResourceResponseItem
The list response item that contains the resource and end points information.

This type is not used in any activity, and only used as part of another schema.

Operation
An operation resource, used to manage asynchronous API requests.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

remove resources zone views (response)

add resources zone views (response)

delete zone views (response)

get zone operations (response)

insert zone views (response)

set service zone views (response)

OperationError
[Output Only] If errors occurred during processing of this operation, this field will be populated.

This type is not used in any activity, and only used as part of another schema.

OperationErrorErrors
[Output Only] The array of errors encountered while processing this operation.

This type is not used in any activity, and only used as part of another schema.

OperationList
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list zone operations (response)

OperationWarnings
[Output Only] If there are issues with this operation, a warning is returned.

This type is not used in any activity, and only used as part of another schema.

OperationWarningsData
[Output only] Metadata for this warning in key:value format.

This type is not used in any activity, and only used as part of another schema.

ResourceView
The resource view object.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get zone views (response)

insert zone views (request)

Resourceviews
Central instance to access all Resourceviews related resource activities

ServiceEndpoint
The service endpoint that may be started in a VM.

This type is not used in any activity, and only used as part of another schema.

ZoneOperationGetMethodBuilder
Retrieves the specified zone-specific operation resource.

ZoneOperationListMethodBuilder
Retrieves the list of operation resources contained within the specified zone.

ZoneOperationMethodsBuilder
A builder providing access to all methods supported on zoneOperation resources. It is not used directly, but through the Resourceviews hub.

ZoneViewAddResourceMethodBuilder
Add resources to the view.

ZoneViewDeleteMethodBuilder
Delete a resource view.

ZoneViewGetMethodBuilder
Get the information of a zonal resource view.

ZoneViewGetServiceMethodBuilder
Get the service information of a resource view or a resource.

ZoneViewInsertMethodBuilder
Create a resource view.

ZoneViewListMethodBuilder
List resource views.

ZoneViewListResourceMethodBuilder
List the resources of the resource view.

ZoneViewMethodsBuilder
A builder providing access to all methods supported on zoneView resources. It is not used directly, but through the Resourceviews hub.

ZoneViewRemoveResourceMethodBuilder
Remove resources from the view.

ZoneViewSetServiceMethodBuilder
Update the service information of a resource view or a resource.

ZoneViewsAddResourcesRequest
The request to add resources to the resource view.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

add resources zone views (request)

ZoneViewsGetServiceResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get service zone views (response)

ZoneViewsList
The response to a list request.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list zone views (response)

ZoneViewsListResourcesResponse
The response to a list resource request.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list resources zone views (response)

ZoneViewsRemoveResourcesRequest
The request to remove resources from the resource view.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

remove resources zone views (request)

ZoneViewsSetServiceRequest
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

set service zone views (request)

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.